// types/graphql/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A (potentially binary) string encoded using base64. */
  Base64String: any;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** A fully qualified reference name (e.g. `refs/heads/master`). */
  GitRefname: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** A string containing HTML code. */
  HTML: any;
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: any;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type Blob = GitObject &
  Node & {
    __typename?: 'Blob';
    /** An abbreviated version of the Git object ID */
    abbreviatedOid: Scalars['String'];
    /** Byte size of Blob object */
    byteSize: Scalars['Int'];
    /** The HTTP path for this Git object */
    commitResourcePath: Scalars['URI'];
    /** The HTTP URL for this Git object */
    commitUrl: Scalars['URI'];
    id: Scalars['ID'];
    /** Indicates whether the Blob is binary or text. Returns null if unable to determine the encoding. */
    isBinary?: Maybe<Scalars['Boolean']>;
    /** Indicates whether the contents is truncated */
    isTruncated: Scalars['Boolean'];
    /** The Git object ID */
    oid: Scalars['GitObjectID'];
    /** The Repository the Git object belongs to */
    repository: Repository;
    /** UTF8 text data or null if the Blob is binary */
    text?: Maybe<Scalars['String']>;
  };
// ...
export type Commit = GitObject &
  Node &
  Subscribable &
  UniformResourceLocatable & {
    __typename?: 'Commit';
    /** An abbreviated version of the Git object ID */
    abbreviatedOid: Scalars['String'];
    /** The number of additions in this commit. */
    additions: Scalars['Int'];
    /**
     * The merged Pull Request that introduced the commit to the repository. If the
     * commit is not present in the default branch, additionally returns open Pull
     * Requests associated with the commit
     */
    associatedPullRequests?: Maybe<PullRequestConnection>;
    /** Authorship details of the commit. */
    author?: Maybe<GitActor>;
    /** Check if the committer and the author match. */
    authoredByCommitter: Scalars['Boolean'];
    /** The datetime when this commit was authored. */
    authoredDate: Scalars['DateTime'];
    /**
     * The list of authors for this commit based on the git author and the Co-authored-by
     * message trailer. The git author will always be first.
     */
    authors: GitActorConnection;
    /** Fetches `git blame` information. */
    blame: Blame;
    /** The number of changed files in this commit. */
    changedFiles: Scalars['Int'];
    /** The check suites associated with a commit. */
    checkSuites?: Maybe<CheckSuiteConnection>;
    /** Comments made on the commit. */
    comments: CommitCommentConnection;
    /** The HTTP path for this Git object */
    commitResourcePath: Scalars['URI'];
    /** The HTTP URL for this Git object */
    commitUrl: Scalars['URI'];
    /** The datetime when this commit was committed. */
    committedDate: Scalars['DateTime'];
    /** Check if committed via GitHub web UI. */
    committedViaWeb: Scalars['Boolean'];
    /** Committer details of the commit. */
    committer?: Maybe<GitActor>;
    /** The number of deletions in this commit. */
    deletions: Scalars['Int'];
    /** The deployments associated with a commit. */
    deployments?: Maybe<DeploymentConnection>;
    /** The tree entry representing the file located at the given path. */
    file?: Maybe<TreeEntry>;
    /** The linear commit history starting from (and including) this commit, in the same order as `git log`. */
    history: CommitHistoryConnection;
    id: Scalars['ID'];
    /** The Git commit message */
    message: Scalars['String'];
    /** The Git commit message body */
    messageBody: Scalars['String'];
    /** The commit message body rendered to HTML. */
    messageBodyHTML: Scalars['HTML'];
    /** The Git commit message headline */
    messageHeadline: Scalars['String'];
    /** The commit message headline rendered to HTML. */
    messageHeadlineHTML: Scalars['HTML'];
    /** The Git object ID */
    oid: Scalars['GitObjectID'];
    /** The organization this commit was made on behalf of. */
    onBehalfOf?: Maybe<Organization>;
    /** The parents of a commit. */
    parents: CommitConnection;
    /** The datetime when this commit was pushed. */
    pushedDate?: Maybe<Scalars['DateTime']>;
    /** The Repository this commit belongs to */
    repository: Repository;
    /** The HTTP path for this commit */
    resourcePath: Scalars['URI'];
    /** Commit signing information, if present. */
    signature?: Maybe<GitSignature>;
    /** Status information for this commit */
    status?: Maybe<Status>;
    /** Check and Status rollup information for this commit. */
    statusCheckRollup?: Maybe<StatusCheckRollup>;
    /** Returns a list of all submodules in this repository as of this Commit parsed from the .gitmodules file. */
    submodules: SubmoduleConnection;
    /**
     * Returns a URL to download a tarball archive for a repository.
     * Note: For private repositories, these links are temporary and expire after five minutes.
     */
    tarballUrl: Scalars['URI'];
    /** Commit's root Tree */
    tree: Tree;
    /** The HTTP path for the tree of this commit */
    treeResourcePath: Scalars['URI'];
    /** The HTTP URL for the tree of this commit */
    treeUrl: Scalars['URI'];
    /** The HTTP URL for this commit */
    url: Scalars['URI'];
    /** Check if the viewer is able to change their subscription status for the repository. */
    viewerCanSubscribe: Scalars['Boolean'];
    /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
    viewerSubscription?: Maybe<SubscriptionState>;
    /**
     * Returns a URL to download a zipball archive for a repository.
     * Note: For private repositories, these links are temporary and expire after five minutes.
     */
    zipballUrl: Scalars['URI'];
  };
// ...
export type CommitEdge = {
  __typename?: 'CommitEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
  /** The item at the end of the edge. */
  node?: Maybe<Commit>;
};

/** The connection type for Commit. */
export type CommitHistoryConnection = {
  __typename?: 'CommitHistoryConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<CommitEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<Commit>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type GitActor = {
  __typename?: 'GitActor';
  /** A URL pointing to the author's public avatar. */
  avatarUrl: Scalars['URI'];
  /** The timestamp of the Git action (authoring or committing). */
  date?: Maybe<Scalars['GitTimestamp']>;
  /** The email in the Git commit. */
  email?: Maybe<Scalars['String']>;
  /** The name in the Git commit. */
  name?: Maybe<Scalars['String']>;
  /** The GitHub user corresponding to the email field. Null if no such user exists. */
  user?: Maybe<User>;
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars['ID'];
};
// ...
export type Organization = Actor &
  MemberStatusable &
  Node &
  PackageOwner &
  ProfileOwner &
  ProjectNextOwner &
  ProjectOwner &
  RepositoryDiscussionAuthor &
  RepositoryDiscussionCommentAuthor &
  RepositoryOwner &
  Sponsorable &
  UniformResourceLocatable & {
// ...
    repository?: Maybe<Repository>;
// ...
  };
// ...
export type PullRequestConnection = {
  __typename?: 'PullRequestConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<PullRequestEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<PullRequest>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type Ref = Node & {
  __typename?: 'Ref';
  /** A list of pull requests with this ref as the head ref. */
  associatedPullRequests: PullRequestConnection;
  /** Branch protection rules for this ref */
  branchProtectionRule?: Maybe<BranchProtectionRule>;
  id: Scalars['ID'];
  /** The ref name. */
  name: Scalars['String'];
  /** The ref's prefix, such as `refs/heads/` or `refs/tags/`. */
  prefix: Scalars['String'];
  /** Branch protection rules that are viewable by non-admins */
  refUpdateRule?: Maybe<RefUpdateRule>;
  /** The repository the ref belongs to. */
  repository: Repository;
  /** The object the ref points to. Returns null when object does not exist. */
  target?: Maybe<GitObject>;
};
// ...
export type Tag = GitObject &
  Node & {
    __typename?: 'Tag';
    /** An abbreviated version of the Git object ID */
    abbreviatedOid: Scalars['String'];
    /** The HTTP path for this Git object */
    commitResourcePath: Scalars['URI'];
    /** The HTTP URL for this Git object */
    commitUrl: Scalars['URI'];
    id: Scalars['ID'];
    /** The Git tag message. */
    message?: Maybe<Scalars['String']>;
    /** The Git tag name. */
    name: Scalars['String'];
    /** The Git object ID */
    oid: Scalars['GitObjectID'];
    /** The Repository the Git object belongs to */
    repository: Repository;
    /** Details about the tag author. */
    tagger?: Maybe<GitActor>;
    /** The Git object the tag points to. */
    target: GitObject;
  };
// ...
export type Tree = GitObject &
  Node & {
    __typename?: 'Tree';
    /** An abbreviated version of the Git object ID */
    abbreviatedOid: Scalars['String'];
    /** The HTTP path for this Git object */
    commitResourcePath: Scalars['URI'];
    /** The HTTP URL for this Git object */
    commitUrl: Scalars['URI'];
    /** A list of tree entries. */
    entries?: Maybe<Array<TreeEntry>>;
    id: Scalars['ID'];
    /** The Git object ID */
    oid: Scalars['GitObjectID'];
    /** The Repository the Git object belongs to */
    repository: Repository;
  };
// ...
export type RepositoryDetailsFragment = {
  __typename?: 'Repository';
  createdAt: any;
  description?: string | null | undefined;
  forkCount: number;
  sshUrl: any;
  name: string;
  owner:
    | { __typename?: 'Organization'; login: string; avatarUrl: any }
    | { __typename?: 'User'; login: string; avatarUrl: any };
  defaultBranchRef?:
    | {
        __typename?: 'Ref';
        target?:
          | { __typename?: 'Blob' }
          | {
              __typename?: 'Commit';
              id: string;
              history: {
                __typename?: 'CommitHistoryConnection';
                edges?:
                  | Array<
                      | {
                          __typename?: 'CommitEdge';
                          node?:
                            | {
                                __typename?: 'Commit';
                                message: string;
                                messageHeadline: string;
                                authoredDate: any;
                                id: string;
                                author?:
                                  | {
                                      __typename?: 'GitActor';
                                      name?: string | null | undefined;
                                      avatarUrl: any;
                                    }
                                  | null
                                  | undefined;
                              }
                            | null
                            | undefined;
                        }
                      | null
                      | undefined
                    >
                  | null
                  | undefined;
              };
            }
          | { __typename?: 'Tag' }
          | { __typename?: 'Tree' }
          | null
          | undefined;
      }
    | null
    | undefined;
};